requirements.txt
# Core Flask
Flask==2.2.5
flask-cors==3.0.10
werkzeug==3.0.0

# Validation
pydantic==1.10.11

# ML/AI
scikit-learn==1.3.2
pandas==2.2.2
joblib==1.3.2

# Database
sqlalchemy==2.0.19

# Configuration
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
flask-limiter==3.5.0

# IBM Cloud SDK Core
ibm-cloud-sdk-core==3.19.0

# IBM Watson Services
ibm-watson==6.1.0

# IBM Cloudant
# ibm-cloudant

# HTTP requests (for watsonx.ai and Orchestrate)
requests==2.31.0

# Utilities
python-dateutil==2.8.2
